+++ src/qubex/simulator/__init__.py
@@ -1,11 +1,13 @@
-from .simulator import Control, MultiControl, Simulator
-from .system import Coupling, System, Transmon
+from .quantum_simulator import Control, MultiControl, QuantumSimulator
+from .quantum_system import Coupling, QuantumSystem, Qubit, Resonator, Transmon
 
 __all__ = [
     "Control",
     "Coupling",
     "MultiControl",
-    "Simulator",
-    "System",
+    "QuantumSimulator",
+    "QuantumSystem",
+    "Qubit",
+    "Resonator",
     "Transmon",
 ]
